fix: move base64 tests to the correct file
